0.16.2 phase A: extract imports/ helpers (factories, http, lm, replay_glue)
First slice of the per-namespace split. `run_wasm_gc/imports.rs` was
1711 lines of dispatch + mixed helper definitions. Cut it down to
the dispatch table (~870 lines) by lifting four orthogonal helper
domains into sibling files:

- `imports/lm.rs` — LM transport (Caller-side `lm_string_to_host`,
  `lm_string_from_host`, `host_print` writer, `val_i64` extractor).
- `imports/replay_glue.rs` — recording/replay state machinery
  (`record_effect_if_recording`, `try_replay`, `json_*` markers).
- `imports/factories.rs` — the `__rt_record_*` builder family
  (`host_*_make`, `host_result_*`, `host_option_*`,
  `host_terminal_size_make`, `host_tcp_connection_*`,
  `host_map_string_list_string_empty`, …).
- `imports/http.rs` — `HttpVerb`, `http_simple_dispatch`,
  `http_body_dispatch`, `http_outcome_*` translators.

`imports.rs` keeps the 29-arm `dispatch_aver_import` match (still
the single source of truth for effect-name → handler binding) plus
re-exports the public surface `decode.rs` consumes (`host_*_make`,
`HttpVerb`, `lm_string_from_host`).

Visibility: helpers exposed as `pub(crate)` so `pub(super) use` in
`imports.rs` can re-export them upward — submodules can't surface
their own items above their parent without crate-wide pub.

No behavioural change. 8/8 record/replay smokes pass, both clippy
invocations clean.

Phase B (next commit): carve the `dispatch_aver_import` arms into
per-namespace `imports/{args,console,disk,env,random,tcp,terminal,
time}.rs` files with `pub(super) fn dispatch(name, …) ->
Result<Option<bool>>`, top-level chains them.
